// featurehub-javascript-client-sdk/app/models/models/sseresult-state.ts
/* tslint:disable */
/* eslint-disable */
/**
 * FeatureServiceApi
 * This describes the API clients use for accessing features
 *
 * The version of the OpenAPI document: 1.1.8
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * error is an inherent state
 * @export
 * @enum {string}
 */
export enum SSEResultState {
    Ack = 'ack',
    Bye = 'bye',
    Failure = 'failure',
    Features = 'features',
    Feature = 'feature',
    DeleteFeature = 'delete_feature',
    Config = 'config',
    Error = 'error'
}